# standard/backend/docs/swagger.yaml
basePath: /api/v1/standard
definitions:
...
  github_com_addp_standard_internal_models.MetricCategory:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      version:
        type: integer
    type: object
...
  github_com_addp_standard_internal_models.UpdateMetricCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      version:
        type: integer
    required:
    - version
    type: object
...
host: localhost:8110
info:
  contact: {}
  title: ADDP Standard API
  version: "1.0"
paths:
...
  /metric-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_addp_standard_internal_models.MetricCategory'
            type: array
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取指标分类列表 | List metric categories
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.metric.read
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_addp_standard_internal_models.MetricCategory'
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建指标分类 | Create metric category
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.metric.create
  /metric-categories/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除指标分类 | Delete metric category
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.metric.delete
    put:
      parameters:
      - description: 更新指标分类 | Update metric category
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_addp_standard_internal_models.UpdateMetricCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 资源版本冲突 | Resource version conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新指标分类 | Update metric category
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.metric.update
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"